+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Unreleased
 
-* Fix `Length::Default` having a layout effect in exact size properties. !!: TODO
+* Fix `Length::Default` having a layout effect in exact size properties.
     - This default value now *disables* the properties, the required behavior for integration with `when` blocks in widgets.
-    - Changed properties: `layout::{size, max_size, width, height, max_width, max_height}`.
+    - Changed properties: `layout::{size, max_size, force_size}` and related properties.
 * Fix buttons at the top-level of `Menu!` not returning focus on click.
 * View-process now tries to guess image format in case of header decode error for an extension or mime defined format.
 * Fix `cargo zng fmt` wrap instability inside macros.